+++ src/scverse_misc/_deprecated.py
@@ -2,10 +2,9 @@ from __future__ import annotations
 
 import inspect
 import sys
-from contextlib import suppress
 from functools import wraps
 from textwrap import indent
-from typing import TYPE_CHECKING, LiteralString
+from typing import TYPE_CHECKING, LiteralString, Protocol, cast
 from warnings import warn
 
 if sys.version_info >= (3, 13):
@@ -29,6 +28,7 @@ class Deprecation(str):
     """
 
     version_deprecated: LiteralString
+    _docmsg: str | None = None
 
     def __new__(cls, version_deprecated: LiteralString, msg: LiteralString = "") -> LiteralString:  # type: ignore[misc]  # typing.Intersection doesn’t exist yet
         if not msg:
@@ -43,7 +43,8 @@ def _deprecated_at[F: Callable[..., object]](
 ) -> Callable[[F], F]:
     """Decorator to indicate that a class, function, or overload is deprecated.
 
-    Wraps :func:`warnings.deprecated` and additionally modifies the docstring to include a deprecation notice.
+    Wraps :func:`warnings.deprecated`. If the scverse_misc Sphinx extension is enabled, the function's documentation will
+    include a deprecation notice.
 
     Args:
         msg: The deprecation message.
@@ -59,24 +60,13 @@ def _deprecated_at[F: Callable[..., object]](
     def decorate(func: F) -> F:
         kind = "function" if func.__name__ == func.__qualname__ else "method"
         warnmsg = f"The {kind} {func.__name__} is deprecated and will be removed in the future"
-
-        doc = inspect.getdoc(func)
-        docmsg = f".. version-deprecated:: {msg.version_deprecated}"
         if len(msg):
-            docmsg += f"\n{indent(msg, 3 * ' ')}"
             warnmsg += f". {msg}" if msg.count("\n") == 0 else f":\n{indent(msg, 4 * ' ')}"
         else:
             warnmsg += "."
-
-        if doc is None:
-            doc = docmsg
-        else:
-            lines = doc.splitlines()
-            body = "\n".join(lines[1:])
-            doc = f"{lines[0]}\n\n{docmsg}\n{body}"
-        func.__doc__ = doc
-
-        return _deprecated(warnmsg, category=category, stacklevel=stacklevel)(func)
+        newmsg = Deprecation(msg.version_deprecated, warnmsg)
+        newmsg._docmsg = str(msg)
+        return _deprecated(newmsg, category=category, stacklevel=stacklevel)(func)
 
     return decorate
 
@@ -87,13 +77,18 @@ else:
     deprecated = _deprecated_at
 
 
-def deprecated_arg[**P, R](
-    arg: LiteralString, msg: Deprecation, *, category: type[Warning] = FutureWarning, stacklevel: int = 1
-) -> Callable[[Callable[P, R]], Callable[P, R]]:
+class CallableWithDeprecatedArg[**P, R](Protocol):
+    __scverse_misc_deprecated_arg__: list[deprecated_arg]
+
+    def __call__(*args: P.args, **kwargs: P.kwargs) -> R: ...
+
+
+class deprecated_arg:
     """Decorator to indicate that a function argument is deprecated.
 
-    Emits a warning when the decorated function is called with the deprecated argument and addtionally modifies the
-    docstring to include a deprecation notice.
+    Emits a warning when the decorated function is called with the deprecated argument.
+    If the `scverse_misc` Sphinx extension is enabled,
+    the documentation will be modified to include a deprecation notice.
 
     Args:
         arg: The deprecated argument.
@@ -107,77 +102,39 @@ def deprecated_arg[**P, R](
         ...     pass
     """
 
-    def decorate(func: Callable[P, R]) -> Callable[P, R]:
-        warnmsg = f"The argument {arg} is deprecated and will be removed in the future."
-        if len(msg):
-            warnmsg += f" {msg}"
+    def __init__(
+        self, arg: LiteralString, msg: Deprecation, *, category: type[Warning] = FutureWarning, stacklevel: int = 1
+    ):
+        self.arg = arg
+        self.msg = msg
+        self.category = category
+        self.stacklevel = stacklevel
 
-        if func.__doc__ is not None:
-            with suppress(ImportError):
-                func.__doc__ = _deprecate_arg_doc(func.__doc__, arg=arg, msg=msg)
+    def __call__[**P, R](self, func: Callable[P, R]) -> CallableWithDeprecatedArg[P, R]:
+        warnmsg = f"The argument {self.arg} is deprecated and will be removed in the future."
+        if len(self.msg):
+            warnmsg += f" {self.msg}"
 
         sig = inspect.signature(func)
-        param = sig.parameters[arg]
+        param = sig.parameters[self.arg]
 
         @wraps(func)
         def wrapped(*args: P.args, **kwargs: P.kwargs) -> R:
             if (
                 param.kind in (inspect.Parameter.KEYWORD_ONLY, inspect.Parameter.POSITIONAL_OR_KEYWORD)
-                and arg in kwargs
+                and self.arg in kwargs
             ):
-                warn(warnmsg, category=category, stacklevel=stacklevel + 1)
+                warn(warnmsg, category=self.category, stacklevel=self.stacklevel + 1)
             else:
                 bound = sig.bind(*args, **kwargs)
-                if arg in bound.arguments and bound.arguments[arg] != param.default:
-                    warn(warnmsg, category=category, stacklevel=stacklevel + 1)
+                if self.arg in bound.arguments and bound.arguments[self.arg] != param.default:
+                    warn(warnmsg, category=self.category, stacklevel=self.stacklevel + 1)
 
             return func(*args, **kwargs)
 
-        return wrapped
+        if not hasattr(func, "__scverse_misc_deprecated_arg__"):
+            func.__scverse_misc_deprecated_arg__ = []  # type: ignore[attr-defined]
+        func.__scverse_misc_deprecated_arg__.append(self)  # type: ignore[attr-defined]
+        wrapped.__scverse_misc_deprecated_arg__ = func.__scverse_misc_deprecated_arg__  # type: ignore[attr-defined]
 
-    return decorate
-
-
-def _deprecate_arg_doc(doc: str, *, arg: str, msg: Deprecation) -> str:
-    from pydocstring import Docstring, Section, SectionKind, Style, emit_google, emit_numpy, parse
-
-    docmsg = f".. version-deprecated:: {msg.version_deprecated}"
-    if len(msg):
-        docmsg += f"\n   {msg}"
-
-    parsed = parse(doc)
-    if parsed.style is Style.PLAIN:
-        return doc
-
-    model = parsed.to_model()
-    if found := next(
-        (
-            (s, section, p, par)
-            for s, section in enumerate(model.sections)
-            if section.kind in {SectionKind.PARAMETERS, SectionKind.KEYWORD_PARAMETERS, SectionKind.OTHER_PARAMETERS}
-            for p, par in enumerate(section.parameters)
-            if arg in par.names
-        ),
-        None,
-    ):
-        s, section, p, par = found
-        if par.description is not None:
-            docmsg += f"\n\n{par.description}"
-        par.description = docmsg
-        params = list(section.parameters)
-        params[p] = par
-        sections = list(model.sections)
-        sections[s] = Section(section.kind, parameters=params)
-        model = Docstring(
-            summary=model.summary,
-            extended_summary=model.extended_summary,
-            deprecation=model.deprecation,
-            sections=sections,
-        )
-    match parsed.style:
-        case Style.GOOGLE:
-            return emit_google(model)
-        case Style.NUMPY:
-            return emit_numpy(model)
-        case _:  # pragma: no cover
-            raise AssertionError
+        return cast(CallableWithDeprecatedArg[P, R], wrapped)